move repo work into modules
