+++ src/store/actions.js
@@ -1,5 +1,6 @@
 import HttpService from "@/services/HttpService";
 import books from "@/store/actions/books"
+import repos from "@/store/actions/repos"
 
 const initialize = ({ commit }) => {
   commit("STOP_LOADING");
@@ -8,30 +9,6 @@ const initialize = ({ commit }) => {
 
 const toggleTheApiVisualizer = ({ commit }) => {
   commit("TOGGLE_API_VISUALIZER");
-};
-
-const clearRepos = ({ commit }) => {
-  commit("SET_REPOS", []);
-};
-
-const loadRepos = ({ commit }) => {
-  commit("START_LOADING");
-  return HttpService.get(`github_repos`, (status, response) => {
-    commit("ADD_API_REQUEST", {
-      url: "GET /api/v1/github_repos",
-      response,
-      status,
-    });
-    commit("SET_REPOS", response.data);
-    commit("SET_CATEGORIES", [
-      ...new Set(
-        response.data.map((repo) => {
-          return repo.category;
-        })
-      ),
-    ]);
-    commit("STOP_LOADING");
-  });
 };
 
 const login = ({ commit }, { username, password, redirectTo = "/" }) => {
@@ -55,7 +32,6 @@ export default {
   initialize,
   login,
   toggleTheApiVisualizer,
-  clearRepos,
-  loadRepos,
   ...books,
+  ...repos,
 };
