+++ tests/unit/store/actions.spec.js
@@ -28,68 +28,6 @@ describe("actions", () => {
     });
   });
 
-  describe("clearRepos", () => {
-    test("initialize calls commit with SET_REPOS", () => {
-      actions.clearRepos({ commit });
-
-      expect(commit).toHaveBeenCalledWith("SET_REPOS", []);
-    });
-  });
-
-  describe("loadRepos", () => {
-    const repos = [
-      {
-        id: 1,
-        name: "vue",
-        full_name: "vuejs/vue",
-        category: "javascript_framework",
-        forks_count: 30644,
-        stargazers_count: 190269,
-        watchers_count: 190269,
-        popularity_rating: 177157,
-      },
-      {
-        id: 2,
-        name: "react",
-        full_name: "facebook/react",
-        category: "javascript_framework",
-        forks_count: 35903,
-        stargazers_count: 177616,
-        watchers_count: 177616,
-        popularity_rating: 168295,
-      },
-    ];
-
-    beforeEach(() => {
-      httpService.get.mockImplementation((_, callback) =>
-        callback(200, { data: repos })
-      );
-      actions.loadRepos({ commit }, "the burbs");
-    });
-
-    test("loadRepos calls commit with START_LOADING", () => {
-      expect(commit).toHaveBeenCalledWith("START_LOADING");
-    });
-    test("loadRepos calls commit with ADD_API_REQUEST", () => {
-      expect(commit).toHaveBeenCalledWith("ADD_API_REQUEST", {
-        url: "GET /api/v1/github_repos",
-        response: { data: repos },
-        status: 200,
-      });
-    });
-    test("loadRepos calls commit with SET_CATEGORIES", () => {
-      expect(commit).toHaveBeenCalledWith("SET_CATEGORIES", [
-        "javascript_framework",
-      ]);
-    });
-    test("loadRepos calls commit with SET_REPOS", () => {
-      expect(commit).toHaveBeenCalledWith("SET_REPOS", repos);
-    });
-    test("loadRepos calls commit with STOP_LOADING", () => {
-      expect(commit).toHaveBeenCalledWith("STOP_LOADING");
-    });
-  });
-
   describe("login", () => {
     describe("when token obtained", () => {
       beforeEach(() => {
@@ -109,11 +47,11 @@ describe("actions", () => {
         actions.login({ commit }, { username: "test", password: "password" });
       });
 
-      test("loadRepos calls commit with START_LOADING", () => {
+      test("login calls commit with START_LOADING", () => {
         expect(commit).toHaveBeenCalledWith("START_LOADING");
       });
 
-      test("loadRepos calls commit with STOP_LOADING", () => {
+      test("login calls commit with STOP_LOADING", () => {
         expect(commit).toHaveBeenCalledWith("STOP_LOADING");
       });
 
@@ -151,77 +89,6 @@ describe("actions", () => {
       test("it doesn't try to set the token if it's not present", () => {
         expect(commit).not.toHaveBeenCalledWith("SET_TOKEN");
       });
-    });
-  });
-
-  describe("removeFavoriteBook", () => {
-    beforeEach(() => {
-      httpService.delete.mockImplementation((_, callback) =>
-        callback(200, {
-          data: {
-            id: 12,
-            user_id: 3,
-            book_id: 2,
-          },
-          errors: [],
-        })
-      );
-      actions.removeFavoriteBook({ commit }, { id: 1 });
-    });
-
-    test("removeFavoriteBook calls commit with ADD_API_REQUEST", () => {
-      expect(commit).toHaveBeenCalledWith("ADD_API_REQUEST", {
-        url: "DELETE /api/v1/user_books/1",
-        response: {
-          data: {
-            book_id: 2,
-            id: 12,
-            user_id: 3,
-          },
-          errors: [],
-        },
-        status: 200,
-      });
-    });
-    test("removeFavoriteBook calls commit with DESTROY_FAVORITE_BOOK", () => {
-      expect(commit).toHaveBeenCalledWith("DESTROY_FAVORITE_BOOK", { id: 1 });
-    });
-  });
-
-  describe("setFavoriteBook", () => {
-    beforeEach(() => {
-      httpService.delete.mockImplementation((_, callback) =>
-        callback(200, {
-          data: {
-            id: 15,
-            user_id: 3,
-            book_id: 2,
-          },
-          errors: [],
-        })
-      );
-      actions.setFavoriteBook({ commit }, { id: 1 });
-    });
-
-    test("setFavoriteBook calls commit with ADD_API_REQUEST", () => {
-      expect(commit).toHaveBeenCalledWith("ADD_API_REQUEST", {
-        url: "POST /api/v1/user_books",
-        response: {
-          data: {
-            token: "imatoken",
-            user: {
-              age: 30,
-              favorites: [],
-              username: "demo_user@test.com",
-            },
-          },
-          errors: [],
-        },
-        status: 200,
-      });
-    });
-    test("setFavoriteBook calls commit with CREATE_FAVORITE_BOOK", () => {
-      expect(commit).toHaveBeenCalledWith("CREATE_FAVORITE_BOOK", { id: 1 });
     });
   });
 });
